+++ test/subform.test.ts
@@ -1,13 +1,6 @@
 import { configure } from "mobx";
-import { getSnapshot, types, applySnapshot, onPatch } from "mobx-state-tree";
-import {
-  Converter,
-  Field,
-  Form,
-  SubForm,
-  RepeatingForm,
-  converters
-} from "../src";
+import { types } from "mobx-state-tree";
+import { Field, Form, SubForm, converters } from "../src";
 
 // "strict" leads to trouble during initialization.
 configure({ enforceActions: true });